+++ src/edgeDriver/src/commands.lua
@@ -5,7 +5,7 @@ local log = require('log')
 local json = require('dkjson')
 local cosock = require "cosock"
 local http = cosock.asyncify "socket.http"
-http.TIMEOUT = 5
+http.TIMEOUT = 15
 local ltn12 = require('ltn12')
 local httpUtil = require('httpUtil')
 local socket = require('socket')
@@ -26,10 +26,12 @@ local updateAvailable = false
 
 --Prevent spamming bad auth info
 local authIsBad = false
+local badAuthCount = 0
+local badAuthThrottleThreshold = 5
 
 --Allow for occasional MyQ errors
 local consecutiveFailureCount = 0
-local consecutiveFailureThreshold = 10
+local consecutiveFailureThreshold = 5
 
 --Handle skipping a refresh iteration if a command was just issued
 local commandIsPending = false
@@ -46,6 +48,10 @@ end
 -- Refresh command
 function command_handler.refresh(driver, callingDevice, skipScan, firstAuth)
 
+  if skipScan == undefined then
+    skipScan = 0
+  end
+
   if commandIsPending == true then
     log.info('Skipping refresh to let command settle.')
     commandIsPending = false
@@ -53,7 +59,14 @@ function command_handler.refresh(driver, callingDevice, skipScan, firstAuth)
   end
 
   if authIsBad == true then
-    log.info('Bad auth.')
+    log.info('Unable to login to MyQ with credentials. Will try again after a delay. Updating credentials will trigger an instant retry.')
+    badAuthCount = badAuthCount + 1
+    if badAuthCount >= badAuthThrottleThreshold then
+      --Reset to allow trying again
+      authIsBad = false
+      badAuthCount = 0
+
+    end
     return
   end
 
@@ -307,23 +320,40 @@ function command_handler.refresh(driver, callingDevice, skipScan, firstAuth)
       myQController:emit_event(myqStatusCap.statusText(newStatus))
     end
 
-  elseif code == 401 and firstAuth == 1 then
-    --Update all devices to show invalid auth
-    local device_list = driver:get_devices() --Grab existing devices
-        for _, device in ipairs(device_list) do
-          device:emit_event(myqStatusCap.statusText('Invalid credentials.'))
-        end
-    authIsBad = true
-    return
+  -- elseif code == 401 then
+  --   --Update all devices to show invalid auth
+  --   local device_list = driver:get_devices() --Grab existing devices
+  --       for _, device in ipairs(device_list) do
+  --         device:emit_event(myqStatusCap.statusText('MyQ login failed'))
+  --       end
+  --   authIsBad = true
+  --   return
 
   else
 
-    --The MyQ API is unreliable. Allow up to 10 failures in a row before we display failure
+    --The MyQ API is unreliable. Allow for a few failures in a row before we display failure
+    log.info('Failure count ' ..consecutiveFailureCount)
     consecutiveFailureCount = consecutiveFailureCount + 1
     if consecutiveFailureCount > consecutiveFailureThreshold then
 
+      --log.info('Server appears down')
       --Update all devices to show server offline
       local offlineStatus = 'Error: MyQ Bridge Offline: ' ..myQController.model
+
+      --401 means the bridge server is up but cannot login to MyQ
+      if code == 401 then
+        log.info(code)
+        offlineStatus = 'Error: Bridge server failed to login to MyQ: ' ..table.concat(res_body)
+        skipScan = 1
+        authIsBad = true
+      end
+
+      --500 errors mean the bridge server is up and logged in but cannot refresh MyQ devices
+      if code == 500 then
+        offlineStatus = 'Error: Bridge server unable to refresh MyQ devices: ' ..table.concat(res_body)
+        skipScan = 1
+      end
+
       local device_list = driver:get_devices() --Grab existing devices
       for _, device in ipairs(device_list) do
         device:offline()
@@ -341,18 +371,32 @@ function command_handler.refresh(driver, callingDevice, skipScan, firstAuth)
           device:emit_event(myqStatusCap.statusText(offlineStatus))
         end
       end
+
+      if (skipScan ~= 1) then
+        local scanReason = ''
+        if code ~= undefined then
+          scanReason = 'last result: ' ..code
+        end
+        log.info('Doing broadcast' ..skipScan)
+        doBroadcast(driver, device, myQController, scanReason)
+      end
+
     end
 
     --If refresh failed with no response at all, try a UDP search to try and auto detect the server (maybe the IP or port changed)
     log.error('Refresh Failed.')
 
-    if (skipScan ~= 1) then
-      doBroadcast(driver, device, myQController)
-    end
+
   end
 end
 
-function doBroadcast(driver, device, myQController)
+function doBroadcast(driver, device, myQController, reason)
+
+  local defaultLookingStatus = 'Searching for bridge server'
+  local currentStatus = myQController:get_latest_state('main', "towertalent27877.myqstatus", "statusText", "unknown")
+  if currentStatus ~= defaultLookingStatus then
+    myQController:emit_event(myqStatusCap.statusText(defaultLookingStatus))
+  end
 
   if driver.server.ip == nil then
     log.info('Refresh: waiting for driver http startup')