+++ src/server/server.js
@@ -12,28 +12,28 @@ var myq; //Holds MyQ connection object
 var myQDeviceMap = {} //Local cache of devices and their statuses
 var searchPending = false;
 var updateAvailable = false;
+var reAuth = true;
 
 const ssdpId = 'urn:SmartThingsCommunity:device:MyQController' //Used in SSDP auto-discovery
 
 
 //Set credentials on myq object (these are always passed-in from calls from the ST hub)
-function myqLogin(email, password){
+async function validateMyqLogin(email, password){
+
+  //Handle missing info
+  if (!email || !password){
+    log('Missing username or password.')
+    return false;
+  }
 
   //If password has been updated, set up new API object
   if (email != myqEmail || password != myqPassword){
-
-    //Handle missing info
-    if (!email || !password){
-      log('Missing username or password.')
-      return false;
-    }
-
-    //Save it
-    log('Got new username/password from hub. Initializing connection to MyQ.')
-    myq = new myQApi.myQApi(email, password);
+    reAuth = true;
+    log('Got new username/password from hub.');
     myqEmail = email;
     myqPassword = password;
   }
+
   return true;
 }
 
@@ -42,15 +42,31 @@ function myqLogin(email, password){
 //Gets devices
 app.post('/devices', async (req, res) => {
   try {
-    if (!myqLogin(req.body.auth.email, req.body.auth.password)){
+
+    let email = req.body.auth.email;
+    let password = req.body.auth.password;
+
+    if (!validateMyqLogin(email, password)){
       return res.sendStatus(401);
     }
 
-    let refreshResult = await myq.refreshDevices();
-    if (!refreshResult){
+    //Connect. Note that the login call automatically calls refresh
+    if (reAuth){
+      myq = new myQApi.myQApi()
+      await myq.login(email, password);
+      reAuth = false;
+    }
+    else{
+      await myq.refreshDevices();
+    }
+
+    if (!myq.accessToken){
+      log(`MyQ login failed`, 1);
+      return res.sendStatus(401);
+    }
+
+    if (myq.apiReturnStatus != 200){
       log(`Refresh failed`, 1);
-      myqEmail = '';
-      myqPassword = '';
       throw new Error('refresh failed');
     }
 
@@ -150,16 +166,16 @@ function startSsdp() {
       if (searchPending || headers.ST != ssdpId || !headers.SERVER_IP || !headers.SERVER_PORT){
         return;
       }
-      searchPending = false;
+      searchPending = true;
       let hubAddress = `http://${headers.SERVER_IP}:${headers.SERVER_PORT}/ping`
       log(`Detected auto-discovery request from SmartThings Hub (${hubAddress}). Replying with bridge server URL.`)
-      let response = await axios.post(hubAddress,
+      await axios.post(hubAddress,
         {
           myqServerPort: port,
           deviceId: headers.DEVICE_ID
         },
         {timeout: 5000})
-      log(`SmartThings hub acknowledged auto-discovery response (status: ${response.status}). If this message repeats, it means the hub cannot connect to this server's URL due to firewall or network issues.`);
+      log(`SmartThings hub acknowledged auto-discovery response. If this message repeats, it means the hub received the bridge server IP/Port but cannot connect to it due to firewall or network issues.`);
     } catch (error) {
         let msg = error.message;
         if (error.response){
@@ -167,6 +183,7 @@ function startSsdp() {
       }
       log(msg, true);
     }
+    searchPending = false;
   });
 }