+++ src/edgeDriver/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2023-09-29T03:09:58.602539965
+
+### Updated
+ - More error message improvements
+
+## 2023-09-12T19:40:09.369702336
+
+### Updated
+ - Improved error handling and logging
+
 ## 2023-09-02T15:52:20.910700768
 
 ### Updated