+++ src/server/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## 1.0.3 (2023-09-06)
+## 1.0.5 (2023-10-11)
+
+### Updated
+ - Update to myq library to 7.5.2
+ - Improve error handling and logging
+
+ ## 1.0.3 (2023-09-06)
 
 ### Updated
  - Updated myQ library